requirements.txt
tensorflow  # Or latest stable version
numpy
pandas
matplotlib
scikit-learn
Pillow
streamlit
